% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPMclean.R
\name{PSPMclean}
\alias{PSPMclean}
\title{Deletes on request all files produced by the PSPManalysis package.}
\usage{
PSPMclean(str = NULL)
}
\arguments{
\item{str}{Character (optional). Only valid argument is 'F'. If not or wrongly
specified the user will be asked whether to do a full clean up or whether to
quit the clean up.}
}
\value{
None.
}
\description{
\code{PSPMclean} deletes all PSPManalysis result files (default) and/or
all executables (hit 'F') in the current directory.
}
\examples{
PSPMclean()

PSPMclean("F")

}
