% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPMhelp.R
\name{PSPMhelp}
\alias{PSPMhelp}
\title{Opens the PSPManalysis manual}
\usage{
PSPMhelp()
}
\value{
None.
}
\description{
\code{PSPMhelp} opens the manual of the the PSPManalysis package in html format.
}
\details{
The manual is created in bookdown format. A PDF version can be downloaded via the 
PDF icon in the menu bar.
}
\examples{
\dontrun{
PSPMhelp()
}

}
