% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showpspm.R
\name{showpspm}
\alias{showpspm}
\title{Shows the model definition file of one of the example models provided with PSPManalysis}
\usage{
showpspm(modelname = NULL)
}
\arguments{
\item{modelname}{(string)
\preformatted{}
              Name of the example model to be displayed.}
}
\description{
\code{showpspm} displays the file contents of one of the physiologically structured
population models that is provided as an example.
}
\details{
showpspm(modelname = NULL)
}
\examples{
\dontrun{
showpspm("Medfly.R")
}

}
