% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSPMhelp.R
\name{PSPMhelp}
\alias{PSPMhelp}
\title{Opens the PSPManalysis manual}
\usage{
PSPMhelp(type = c("html", "pdf"))
}
\arguments{
\item{type}{String (optional). Either "pdf" or "html". Default if "html"}
}
\value{
None.
}
\description{
\code{PSPMhelp} opens the manual of the the PSPManalysis package in either html or pdf format.
}
\examples{
PSPMhelp()

PSPMhelp("pdf")

}
